release

+++ app/Http/Controllers/VacancyController.php
@@ -7,10 +7,16 @@ use App\Models\Vacancy;
 use App\Models\City;
 use App\Models\WorkType;
 use App\Models\Skill;
+use App\Models\WorkSchedule;
+use App\Models\Experience;
+use App\Models\Education;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class VacancyController extends Controller
 {
+    use AuthorizesRequests;
+
     public function show($id)
     {
         $vacancy = Vacancy::findOrFail($id);
@@ -22,7 +28,11 @@ class VacancyController extends Controller
         $cities = City::all();
         $workTypes = WorkType::all();
         $skills = Skill::all();
-        return view('vacancies.create', compact('cities', 'workTypes', 'skills'));
+        $workSchedules = WorkSchedule::all();
+        $experiences = Experience::all();
+        $education = Education::all();
+
+        return view('vacancies.create', compact('cities', 'workTypes', 'skills', 'workSchedules', 'experiences', 'education'));
     }
 
     public function store(VacancyRequest $request)
@@ -41,15 +51,25 @@ class VacancyController extends Controller
     public function edit($id)
     {
         $vacancy = Vacancy::findOrFail($id);
+
+        $this->authorize('edit', $vacancy->user);
+
         $cities = City::all();
         $workTypes = WorkType::all();
         $skills = Skill::all();
-        return view('vacancies.edit', compact('vacancy', 'cities', 'workTypes', 'skills'));
+        $workSchedules = WorkSchedule::all();
+        $experiences = Experience::all();
+        $education = Education::all();
+
+        return view('vacancies.edit', compact('vacancy', 'cities', 'workTypes', 'skills', 'workSchedules', 'experiences', 'education'));
     }
 
     public function update(VacancyRequest $request, $id)
     {
         $vacancy = Vacancy::findOrFail($id);
+
+        $this->authorize('edit', $vacancy->user);
+
         $vacancy->update($request->validated());
 
         if ($request->has('skills')) {
@@ -64,6 +84,9 @@ class VacancyController extends Controller
     public function destroy($id)
     {
         $vacancy = Vacancy::findOrFail($id);
+
+        $this->authorize('edit', $vacancy->user);
+
         $vacancy->delete();
 
         return redirect()->route('home');
